cleaned mapping code
