+++ packages/admin-client/src/types/mapping-form.ts
@@ -1,0 +1,6 @@
+export type MappingFormData = {
+    url: string;
+    modelName: string;
+    historyComponentID: string;
+    provider: string;
+  };
