--- packages/admin-ui/src/types/langflow-mapping.ts
@@ -1,5 +1,0 @@
-export interface LangFlowMapping {
-    model: string;
-    url: string;
-    historyComponentID: string;
-  }
